feat: cleaned index.js and updated notes_controller and user_controller with right routes. Updated test to work properly in notes too
