+++ models/Note.js
@@ -15,6 +15,7 @@ noteSchema.set('toJSON', {
         returnedObject.id = returnedObject._id;
         delete returnedObject._id;
         delete returnedObject._v;
+        delete returnedObject.__v;
     },
 });
