+++ package.json
@@ -5,7 +5,7 @@
     "scripts": {
         "dev": "cross-env NODE_ENV=development nodemon index.js",
         "start": "cross-env NODE_ENV=production node index.js",
-        "test": "cross-env NODE_ENV=test jest --verbose --silent tests/user.test.js",
+        "test": "cross-env NODE_ENV=test jest --verbose --silent tests/",
         "test:watch": "npm run test -- --watch"
     },
     "author": "",
