+++ requests/delete_note.rest
@@ -1,1 +1,1 @@
-DELETE http://localhost:3001/api/notes/6866b9f8cf0d10915d77e70e
+DELETE http://localhost:3001/api/notes/686fb0185718c15530b78479
