+++ requests/get_note_by_id.rest
@@ -1,1 +1,1 @@
-GET http://localhost:3001/api/notes/1
+GET http://localhost:3001/api/notes/686fb0185718c15530b78479
