+++ requests/put_note.rest
@@ -1,4 +1,4 @@
-PUT http://localhost:3001/api/notes/6866acc036a9325c7fb218d1
+PUT http://localhost:3001/api/notes/686fb0185718c15530b78479
 Content-Type: application/json
 
 {
